Allow Boolean options in environment variables.
