+++ autopush/settings.py
@@ -11,22 +11,9 @@ from autopush.db import (
     Router
 )
 from autopush.metrics import DatadogMetrics, TwistedMetrics
+from autopush.utils import canonical_url
 
 from autopush.pinger.pinger import Pinger
-
-
-default_ports = {
-    "ws": 80,
-    "http": 80,
-    "wss": 443,
-    "https": 443,
-}
-
-
-def canonical_url(scheme, hostname, port=None):
-    if port is None or port == default_ports.get(scheme):
-        return "%s://%s" % (scheme, hostname)
-    return "%s://%s:%s" % (scheme, hostname, port)
 
 
 class MetricSink(object):
